requirements.txt
flask
python-dotenv
resend
gunicorn
